Added bayes_opt optimisation plugin
