+++ src/vai_lab/examples/xml_files/bayes_opt_demo.xml
@@ -1,0 +1,36 @@
+<pipeline>
+    <Initialiser name="Initialiser">
+        <relationships>
+            <child name="Decision Making" />
+        </relationships>
+        <coordinates>
+            [(350.0,50),0,{}]
+        </coordinates>
+    </Initialiser>
+    <DecisionMaking name="Decision Making">
+        <relationships>
+            <parent name="Initialiser" />
+            <child name="Output" />
+        </relationships>
+        <coordinates>
+            [(350.0,350.0),2,{0:'d0-u2'}]
+        </coordinates>
+        <plugin type="bayes_opt">
+            <f>
+                 None
+            </f>
+            <pbounds>
+                 {'CsPbI': (0,1), 'MAPbI': (0, 1), 'FAPbI': (0, 1)}
+            </pbounds>
+        </plugin>
+    </DecisionMaking>
+    <Output name="Output">
+        <plugin type="" />
+        <relationships>
+            <parent name="Decision Making" />
+        </relationships>
+        <coordinates>
+            [(350.0,650),1,{2:'d2-u1'}]
+        </coordinates>
+    </Output>
+</pipeline>
